Add files via upload
Hi there, i managed to fix my script, and now i added the Blood Hunter class made by Matthew Mercer. I know that's not an "official" class yet, but it's largely used by the player base, and it's one of my favourite classes too.
i put my script in the ranger character file and features file, hoping it would work like the revised range, and it does. I modified the__init__.py script too in order to make the program acknowledge my class. I don't know if that was the right thing to do, but as long as it works for me i hope is fine.
I would love to see my script added in the main project, but i made this project to practice with python since I'm a newbie , so have fun using it to whoever needs it (i hope it works for you too).
It's a fun class to play, and a great addiction to your party of adventurers.
Thanks for the attention ad have a nice day.