added spell description to appear for race spells
The spellsheets shouldn't be restricted to spellcasters. I suppose if a char gains access to a spell by a feature or any other means, the spellsheet should include it. So I changed a line in `make_sheets` so it calls for the proper method in case the char has any spell listed. 
Also, small changes were made to equipment_reader.py