MSavage latex template: Get rid of spurious indents
Text fields in the MSavage latex template look nice if the
first line is not indented. However, if text input in the
MSavage form begins with a newline, then this leads the
dndtemplate itself to indent the first line. This patch
gets rid of a couple of newlines, and in so doing gets
rid of some indented text fields.

Furthermore, some jinja [% if ... %] declarations also
lead to newlines, when certain variables are undefined.
Add minus-signs in those definitions to strip unwanted
whitespace