MSavage latex template: Proficiency markers
Recent changes the MSavage dndtemplate treat proficiency markers
differently. You can now set half proficiency (e.g., from Jack
of all Traits), normal proficiency, and expertise. To that end,
syntax for marking proficiencies has been changed from boolean
(True or False) to numeral (0.5 for Jack of all Traits, 1 for
normal proficiency, and 2 for expertise). This patch sets
proficiency to 1 if proficiency is detected, so that proficiency
is marked on the character sheet.

Note that the dndtemplate also allows for a specific marker for
half proficiency, which looks really great, but that would mean
that we carry those stats in dungeonsheets as well. Also,
dndtemplate now is able to calculate quite a number of modifiers
as well, but we already do so in dungeonsheets, and I don't see
much point in doing this twice.