MSavage latex template: Make full use of proficiency markers
On second thought, contrary to what I claimed in my previous
patch, it actually is not all that hard to find out whether
a character is proficient, has jack of all trades, has expertise
or has remarkable athlete. So I amended the template a little
bit further, so that skills with half proficiency (jack of all
trades and remarkable athlete) are now explicitly marked as such.