MSavage latex template: Prevent spurious newlines
The functions in character.py that format text for
additional feats and equipment are slightly
overzealous in splitting strings on newlines. This results
in spurious newlines in the equipment box and the features
box on the first page of the character sheet.

We now only add a newline on the character sheet
if the input contains two newlines, and not just one. This
issue was especially visible when using, for instance, the
explorer_pack as defined in the equipment reader, which itself
contains various newlines that ought not be displayed
in the equipment box.