MSavage latex template: Implement halfcaster spellsheet
This patch implements the halfcaster spellsheet for the
MSavage latex template. As with the fillable PDF, select the
halfcaster spellsheet only when the character has no spells of
level 6 and higher, _and_ the character's spell list overflows
the number of slots for one of the spell level lists on the
character's fullcaster spellsheet.