MSavage latex template: Add proficiences listed in proficiencies_text
The MSavage latex template loads "Other proficiencies" from a character
string named chosen_tools. However, this string is not documented in
the character template, which instead lists the proficiencies_text
list variable as the main way to add proficiencies such as thieves'
tools and musical instruments. As a consequence, these proficiencies
are omitted in the MSavage latex template.

This patch copies the proficiencies listed in proficiencies_text
over to chosen_tools, so that they are added to the MSavage latex
template. It does so conditionally, that is, only if chosen_tools
is empty, in order to maintain backwards compatibility.