MSavage latex template: More room for equipment
The equipment box on the MSavage latex template still
quite easily overflows. Previously, this was in part
due to the fact that weight and carrying capacity both
occupied a single line in the equipment box. This patch
puts weigt and carrying capacity on a single line by
displaying it in a smaller font size and by rounding
weight to whole numbers instead of two decimals.