character.py: Fix adding subclasses
For every class, character.py will add subclass=(None)
when initially subclasses==0. However, this approach
later fails an assertion that checks whether the list
of subclasses is equally long as the list of classes.
Specifically, this is problematic for multiclass
characters that have one class with a subclass
already defined by the player (so that subclasses is
not 0), and another class below the level that adds
a subclass (e.g., Bard 5 + Ranger 1). Such
characters have len(subclasses) == 1 and len(classlist)
==2.

This patch changes the rules for adding subclass=(None).
We now check whether the number of subclasses defined is
shorter than the length of classes, and if so, we add
subclass=(None) for the difference.

This is a bugfix (no issue reported).